+++ engine/src/main/scala/cromwell/server/CromwellRootActor.scala
@@ -206,12 +206,24 @@ abstract class CromwellRootActor(terminator: CromwellTerminator,
   lazy val executionTokenLogInterval: Option[FiniteDuration] =
     systemConfig.as[Option[Int]]("hog-safety.token-log-interval-seconds").map(_.seconds)
 
+  private lazy val quotaExhaustionJobControlEnabled: Boolean =
+    systemConfig.as[Option[Boolean]]("quota-exhaustion-job-start-control.enabled").getOrElse(false)
+  private lazy val quotaExhaustionThresholdInMins: Long =
+    systemConfig.as[Option[Long]]("quota-exhaustion-job-start-control.threshold-minutes").getOrElse(15)
+  private lazy val groupMetricsActor: ActorRef =
+    context.actorOf(
+      GroupMetricsActor.props(EngineServicesStore.engineDatabaseInterface, quotaExhaustionThresholdInMins)
+    )
+  private lazy val groupMetricsActorForJTDA: Option[ActorRef] =
+    if (quotaExhaustionJobControlEnabled) Option(groupMetricsActor) else None
+
   lazy val jobRestartCheckTokenDispenserActor: ActorRef = context.actorOf(
     JobTokenDispenserActor.props(serviceRegistryActor,
                                  jobRestartCheckRate,
                                  restartCheckTokenLogInterval,
                                  "restart checking",
-                                 "CheckingRestart"
+                                 "CheckingRestart",
+                                 groupMetricsActorForJTDA
     ),
     "JobRestartCheckTokenDispenser"
   )
@@ -220,13 +232,11 @@ abstract class CromwellRootActor(terminator: CromwellTerminator,
                                  jobExecutionRate,
                                  executionTokenLogInterval,
                                  "execution",
-                                 ExecutionStatus.Running.toString
+                                 ExecutionStatus.Running.toString,
+                                 groupMetricsActorForJTDA
     ),
     "JobExecutionTokenDispenser"
   )
-
-  lazy val groupMetricsActor: ActorRef =
-    context.actorOf(GroupMetricsActor.props(EngineServicesStore.engineDatabaseInterface))
 
   lazy val workflowManagerActor = context.actorOf(
     WorkflowManagerActor.props(